+++ types/markdown.d.ts
@@ -1,6 +1,6 @@
 declare module '*.md' {
   import { Serializable } from 'lib/utils/serialization'
-  import { RenderablePost } from 'lib/pages/posts/[slug]'
+  import { RenderablePost } from 'lib/pages/posts/[slug]/server'
 
   const content: Serializable<RenderablePost>
   export default content
